+++ CMakeLists.txt
@@ -15,7 +15,7 @@ include_directories(src/include duckdb/third_party/httplib duckdb/extension/auto
 set(EXTENSION_SOURCES
         src/quack-startstop.cpp   src/quack_extension.cpp   src/quack_log.cpp         src/quack_server.cpp
         src/quack_catalog.cpp     src/quack_http_server.cpp src/quack_message.cpp     src/quack_storage.cpp
-        src/quack_client.cpp      src/quack_insert.cpp      src/quack_scan.cpp
+        src/quack_client.cpp      src/quack_insert.cpp      src/quack_scan.cpp src/quack_uri.cpp
 )
 
 build_static_extension(${TARGET_NAME} ${EXTENSION_SOURCES})
