+++ test/sql/attach_ctas.test
@@ -16,7 +16,7 @@ set rpc_default_token='asdf'
 
 
 statement ok con1
-call rpc_start({server});
+call quack_serve({server});
 
 statement ok con2
 attach {server} as rpc
