+++ test/sql/secret.test
@@ -11,7 +11,7 @@ statement ok con1
 set rpc_default_token='s3kr3t'
 
 statement ok con1
-call rpc_start('quack:localhost:20101');
+call quack_serve('quack:localhost:20101');
 
 sleep 100 millisecond
