+++ test/sql/secret_rpc_call.test
@@ -1,5 +1,5 @@
 # name: test/sql/secret_rpc_call.test
-# description: verify rpc_call(uri, sql) picks up auth token from quack secrets
+# description: verify quack_query(uri, sql) picks up auth token from quack secrets
 # group: [sql]
 
 require quack
@@ -11,30 +11,30 @@ statement ok con1
 set rpc_default_token='s3kr3t'
 
 statement ok con1
-call rpc_start('quack:localhost:20102');
+call quack_serve('quack:localhost:20102');
 
 sleep 100 millisecond
 
-# --- 1. Matching secret: rpc_call succeeds with the secret's token ---
+# --- 1. Matching secret: quack_query succeeds with the secret's token ---
 
 statement ok con2
 create secret rpc_ok (type quack, token 's3kr3t', scope 'quack:localhost:20102')
 
 query II con2
-from rpc_call('quack:localhost:20102', 'select 1, 2')
+from quack_query('quack:localhost:20102', 'select 1, 2')
 ----
 1	2
 
 statement ok con2
 drop secret rpc_ok
 
-# --- 2. Wrong secret: rpc_call fails with Authentication failed ---
+# --- 2. Wrong secret: quack_query fails with Authentication failed ---
 
 statement ok con2
 create secret rpc_bad (type quack, token 'nope', scope 'quack:localhost:20102')
 
 statement error con2
-from rpc_call('quack:localhost:20102', 'select 1')
+from quack_query('quack:localhost:20102', 'select 1')
 ----
 Authentication failed
 
@@ -47,7 +47,7 @@ statement ok con2
 set rpc_default_token='s3kr3t'
 
 query I con2
-from rpc_call('quack:localhost:20102', 'select 7')
+from quack_query('quack:localhost:20102', 'select 7')
 ----
 7
